refactor(local-books): move createMemoryTokenStore to test helpers
The in-memory token store is a test double with no production caller
(only sync.test.ts used it). Move it out of src/token-store.ts into
test/helpers.ts so the shipped module carries only the file store.
